+++ sdk/keyvault/azure_security_keyvault_secrets/tests/secret_client.rs
@@ -41,7 +41,7 @@ async fn secret_roundtrip(ctx: TestContext) -> Result<()> {
     let secret = client
         .set_secret("secret-roundtrip", body.try_into()?, None)
         .await?
-        .into_body()?;
+        .into_model()?;
     assert_eq!(secret.value, Some("secret-value".into()));
 
     // Get a specific version of a secret.
@@ -55,7 +55,7 @@ async fn secret_roundtrip(ctx: TestContext) -> Result<()> {
             }),
         )
         .await?
-        .into_body()?;
+        .into_model()?;
     assert_eq!(secret.value, Some("secret-value".into()));
 
     Ok(())
@@ -82,7 +82,7 @@ async fn update_secret_properties(ctx: TestContext) -> Result<()> {
     let secret = client
         .set_secret("update-secret", body.try_into()?, None)
         .await?
-        .into_body()?;
+        .into_model()?;
     assert_eq!(secret.value, Some("secret-value".into()));
 
     // Update secret properties.
@@ -97,7 +97,7 @@ async fn update_secret_properties(ctx: TestContext) -> Result<()> {
     let secret = client
         .update_secret_properties("update-secret", properties.try_into()?, None)
         .await?
-        .into_body()?;
+        .into_model()?;
     assert_eq!(secret.content_type, Some("text/plain".into()));
     assert_eq!(
         secret.tags.expect("expected tags").get("test-name"),
@@ -131,7 +131,7 @@ async fn list_secrets(ctx: TestContext) -> Result<()> {
             None,
         )
         .await?
-        .into_body()?;
+        .into_model()?;
     assert_eq!(secret1.value, Some("secret-value-1".into()));
 
     let secret2 = client
@@ -141,7 +141,7 @@ async fn list_secrets(ctx: TestContext) -> Result<()> {
             None,
         )
         .await?
-        .into_body()?;
+        .into_model()?;
     assert_eq!(secret2.value, Some("secret-value-2".into()));
 
     // List secrets.
@@ -183,7 +183,7 @@ async fn purge_secret(ctx: TestContext) -> Result<()> {
             None,
         )
         .await?
-        .into_body()?;
+        .into_model()?;
 
     // Delete the secret.
     let name = secret.resource_id()?.name;
@@ -247,7 +247,7 @@ async fn round_trip_secret_verify_telemetry(ctx: TestContext) -> Result<()> {
                 let secret = client
                     .set_secret("secret-roundtrip-instrument", body.try_into()?, None)
                     .await?
-                    .into_body()?;
+                    .into_model()?;
                 assert_eq!(secret.value, Some("secret-value-instrument".into()));
 
                 // Get a specific version of a secret.
@@ -261,7 +261,7 @@ async fn round_trip_secret_verify_telemetry(ctx: TestContext) -> Result<()> {
                         }),
                     )
                     .await?
-                    .into_body()?;
+                    .into_model()?;
                 assert_eq!(secret.value, Some("secret-value-instrument".into()));
                 Ok(())
             })